third_party/WebKit/LayoutTests/fast/js/Promise-then-in-workers-expected.txt

[Worker] Test Promise.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Starting worker: resources/Promise-then-in-workers.js
PASS [Worker] thisInInit === firstPromise is false
PASS [Worker] thisInInit === global is true
PASS [Worker] firstPromise instanceof Promise is true
PASS [Worker] secondPromise instanceof Promise is true
PASS [Worker] firstPromise.then(null) threw exception TypeError: onFulfilled must be a function or undefined.
PASS [Worker] firstPromise.then(undefined, null) threw exception TypeError: onRejected must be a function or undefined.
PASS [Worker] firstPromise.then(37) threw exception TypeError: onFulfilled must be a function or undefined.
PASS [Worker] thisInFulfillCallback === firstPromise is false
PASS [Worker] thisInFulfillCallback === secondPromise is false
PASS [Worker] thisInFulfillCallback === global is true
PASS [Worker] result is "hello"
PASS successfullyParsed is true

TEST COMPLETE
